persist theme preference
